refactor(led): separate led and led_bar
