+++ app/model_car/CMakeLists.txt
@@ -54,6 +54,7 @@ target_sources(${CMAKE_PROJECT_NAME} PRIVATE
 
     interfaces/rc.c
     interfaces/led.c
+    interfaces/led_bar.c
     interfaces/flash.c
     interfaces/config.c
