+++ app/model_car/interfaces/interface.h
@@ -29,7 +29,7 @@ void led_chasis_set(bool en);
 bool led_chasis_get(void);
 void led_fire_set(bool en);
 bool led_fire_get(void);
-void led_string_set(bool en);
+void led_bar_set(bool en);
 void thro_led_update(int16_t thro);
 void thro_led_brake(void);
 void thro_led_set(uint8_t r, uint8_t g, uint8_t b);
